feat: initialize navigation view composer
