+++ app/Providers/AppServiceProvider.php
@@ -2,13 +2,11 @@
 
 namespace App\Providers;
 
-use App\View\Composers\MenuComposer;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\File;
+use App\View\Composers\NavigationsComposer;
+use App\View\Composers\TranslationsComposer;
 use Illuminate\Support\Facades\Mail;
-use Illuminate\Support\Facades\View as ViewFacade;
+use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\View\View;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -41,22 +39,8 @@ class AppServiceProvider extends ServiceProvider
             Mail::alwaysTo($devMail);
         }
 
-        ViewFacade::composer(['layouts.app'], MenuComposer::class);
+        View::composer('app', NavigationsComposer::class);
 
-        ViewFacade::composer('app', function (View $view) {
-            $locales = [];
-
-            foreach (File::directories(\resource_path('lang')) as $dir) {
-                $trans = [];
-
-                foreach (File::files($dir) as $file) {
-                    $trans[\basename($file, '.php')] = require $file;
-                }
-
-                $locales[\basename($dir)] = Arr::dot($trans);
-            }
-
-            $view->with('translations', $locales);
-        });
+        View::composer('app', TranslationsComposer::class);
     }
 }
